// src/domains/MoMo/types/getPaymentTransaction.d.ts
// Generated by https://quicktype.io

export interface GetPaymentTransaction {
    partnerCode: string;
    requestId: string;
    orderId: string;
    extraData: string;
    amount: number;
    transId: number;
    payType: '' | 'web' | 'qr';
    resultCode: number;
    refundTrans: any[];
    responseTime: number;
    message: string;
    promotionInfo: any[];
}